Automate the process of running a notebook
When exiting the program, the container launched with Docker is now
stopped and removed automatically. Also, a call of 'runjnb.sh' as root
will cause the process to terminate.

In the practice notebook some minor changes were made: increased the
number of elements in a mesh and tolerance for accepting boundary points
as source ones.